// angular/src/app/core/models/warranty/claim.model.ts
import { BaseModel } from '../base.model';

export interface ClaimModel extends BaseModel {
  processFlagLabel?: string;
  dlrName?: string;
  reasonCode?: string;
  flatratedlr?: string;
  pwrdlr?: string;
  dealerName?: boolean;
  distCode?: boolean;
  sfx?: string;
  twcNo?: string;
  sTwcNo?: string;
  claimantcode?: string;
  invoiceNo?: string;
  dealerCode?: string;
  kmFlag?: string;
  dealerClaimNo?: string;
  processFlag?: string;
  warrantyType?: string;
  franchise?: string;
  cmId?: string;
  warrantyAppCode?: string;
  nvFlag?: string;
  fvFlag?: string;
  wmi?: string;
  vds?: string;
  cd?: string;
  vis?: string;
  orderno?: string;
  lineOffDate: any;
  deliveryDate?: any;
  repairDate?: any;
  odometer?: string;
  authoType?: string;
  authoNo?: string;
  dataId?: string;
  tc?: string;
  subletDesc?: string;
  t1Code?: string;
  t2Code?: string;
  t2CodeType?: string;
  condition?: string;
  cause?: string;
  remedy?: string;
  ofpLocalFlag?: string;
  ofpNo?: string;
  pwrDlr?: string;
  prrDlr?: string;
  flatRateDrl?: string;
  freePm?: string;
  dlrComment?: string;
  laborAdj?: number;
  partsAdj?: number;
  subletAdj?: number;
  distComment?: string;
  roState?: string;
  sumSubletSubTotal?: number;
  sumLaborSubTotal?: number;
  laborTotalHour?: number;
  sumPartsSubTotal?: number;
  partsTotal?: number;
  total?: number;
  currentRecord?: number;
  totalRecord?: number;
  warrantyClaimLaborDTOs?: Array<any>;
  warrantyClaimSubletDTOs?: Array<any>;
  warrantyClaimPartsDTOs?: Array<any>;
  t3Codes: Array<any>;
  state?: object;
  tmvState?: object;
  currency: string;
  currencyLabel: string;
}